Extract Models and Abstraction to Core namespace. Implemented dependency inversion.
