+++ Core/IVehicleRepository.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
-using VegaSPA.Models;
+using VegaSPA.Core.Models;
 
-namespace VegaSPA.Data
+namespace VegaSPA.Core
 {
     public interface IVehicleRepository : IRepository<Vehicle>
     {
