+++ Core/Models/Vehicle.cs
@@ -4,7 +4,7 @@ using System.Collections.ObjectModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VegaSPA.Models
+namespace VegaSPA.Core.Models
 {
     // TODO: Identity jump 1000, identity behaviour
     [Table("Vehicles")]
